+++ hyracks-fullstack/hyracks/hyracks-dataflow-std/src/main/java/org/apache/hyracks/dataflow/std/join/InMemoryHashJoinOperatorDescriptor.java
@@ -155,7 +155,7 @@ public class InMemoryHashJoinOperatorDescriptor extends AbstractOperatorDescript
                 nullWriters1 = null;
             }
 
-            final int memSizeInBytes = memSizeInFrames * jobletCtx.getInitialFrameSize();
+            final long memSizeInBytes = (long) memSizeInFrames * (long) jobletCtx.getInitialFrameSize();
             final IDeallocatableFramePool framePool = new DeallocatableFramePool(jobletCtx, memSizeInBytes);
             final ISimpleFrameBufferManager bufferManager = new FramePoolBackedFrameBufferManager(framePool);
