+++ hyracks-fullstack/hyracks/hyracks-storage-am-lsm-invertedindex/src/main/java/org/apache/hyracks/storage/am/lsm/invertedindex/dataflow/LSMInvertedIndexSearchOperatorNodePushable.java
@@ -82,7 +82,7 @@ public class LSMInvertedIndexSearchOperatorNodePushable extends IndexSearchOpera
         }
         this.numOfFields = numOfFields;
         // Intermediate and final search result will use this buffer manager to get frames.
-        this.framePool = new DeallocatableFramePool(ctx, frameLimit * ctx.getInitialFrameSize());
+        this.framePool = new DeallocatableFramePool(ctx, (long) frameLimit * (long) ctx.getInitialFrameSize());
         this.bufferManagerForSearch = new FramePoolBackedFrameBufferManager(framePool);
         // Keep the buffer manager in the hyracks context so that the search process can get it via the context.
         TaskUtil.put(HyracksConstants.INVERTED_INDEX_SEARCH_FRAME_MANAGER, bufferManagerForSearch, ctx);
